# Optimize compiler phase assembly