--- a/test/benchmarks/src/main/scala/scala/tools/nsc/PhaseAssemblyBenchmark.scala
+++ b/test/benchmarks/src/main/scala/scala/tools/nsc/PhaseAssemblyBenchmark.scala
@@ -19,7 +19,7 @@ class PhaseAssemblyBenchmark {
   var data: Data[_] = _
 
   @Param(Array("1", "4", "8", "16"))
-  var size: Int = 0
+  var size: Int = 16
 
   @Setup
   def setup(): Unit = {
@@ -47,7 +47,8 @@ class PhaseAssemblyBenchmark {
     val g = s.global
     val graph = g.phasesSetToDepGraph(s.components.reverse)
     graph.removeDanglingNodes()
-    graph.collapseHardLinksAndLevels(graph.getNodeByPhase("parser"), 1)
+    graph.collapseHardLinks()
+    graph.assignLevelsAndDetectCycles(graph.getNodeByPhase("parser"))
     graph
   }
 }