I noticed in profiles that filtering the the `before` set
for hard links generated a lot of garbage in compilers configured
with a large number of plugins.

This commit filters on the fly instead.

Fixes scala/scala-dev#538