@som-snytt 2019-08-01
I'm surprised to hear someone has tricked out their compiler so that they noticed. Or was this just a test scenario?