@som-snytt 2019-08-01
The other speculation is at https://github.com/scala/scala/pull/8297

I noticed that the final sort order is by "level" and "name", where the name is the first phase name of a group bound by "runsRightAfter".

So that is why fields runs before tailcalls.

> Node(tailcalls) @ 9 < HashSet() > HashSet(SoftEdge(tailcalls->uncurry)) = Some(List(scala.tools.nsc.Global$tailCalls$@fbaecf85, scala.tools.nsc.Global$specializeTypes$@c48471fb))
> Node(fields) @ 9 < HashSet(SoftEdge(explicitouter->fields)) > HashSet(SoftEdge(fields->uncurry)) = Some(List(scala.tools.nsc.Global$fields$@b4057eb9))

There is a code comment for fields to run before erasure, but it doesn't say runsBefore, which is unused. I wonder why.

One of the partests defines a plugin "ploogin" which happens to precede "terminal" also for alphabetical reasons. This is not reassuring. I wonder if it should warn that the order is arbitrary. Usually you really do care about phase order; not caring is an edge case. (Oh, phase assembly pun!)