@diesalbla 2019-08-02
@retronym Would this be good enough to satisfy the needs of https://github.com/scala/scala-dev/issues/538 /