@som-snytt 2019-08-02
Glancing at the linked ticket and its linked WIP, I see that we forgot to ask for a test. And a benchmark. I can't tell if this is deadpan humo(u)r.