@retronym 2019-08-06
I've pushed a benchmark that shows an improvement under this PR.

## Before

```
[info] Benchmark                                                         Mode  Cnt        Score        Error   Units
[info] PhaseAssemblyBenchmark.assemble                                   avgt   10  1832344.612 ±  56469.975   ns/op
[info] PhaseAssemblyBenchmark.assemble:·gc.alloc.rate                    avgt   10      869.791 ±     27.415  MB/sec
[info] PhaseAssemblyBenchmark.assemble:·gc.alloc.rate.norm               avgt   10  2507836.552 ±     36.410    B/op
[info] PhaseAssemblyBenchmark.assemble:·gc.churn.PS_Eden_Space           avgt   10      875.816 ±    156.800  MB/sec
[info] PhaseAssemblyBenchmark.assemble:·gc.churn.PS_Eden_Space.norm      avgt   10  2523822.122 ± 424947.103    B/op
[info] PhaseAssemblyBenchmark.assemble:·gc.churn.PS_Survivor_Space       avgt   10        0.052 ±      0.074  MB/sec
[info] PhaseAssemblyBenchmark.assemble:·gc.churn.PS_Survivor_Space.norm  avgt   10      150.541 ±    220.173    B/op
[info] PhaseAssemblyBenchmark.assemble:·gc.count                         avgt   10       20.000               counts
[info] PhaseAssemblyBenchmark.assemble:·gc.time                          avgt   10       19.000                   ms
```

## After
```
[info]
[info] Benchmark                                                         Mode  Cnt       Score        Error   Units
[info] PhaseAssemblyBenchmark.assemble                                   avgt   10  908735.258 ±  25198.595   ns/op
[info] PhaseAssemblyBenchmark.assemble:·gc.alloc.rate                    avgt   10     253.003 ±      7.018  MB/sec
[info] PhaseAssemblyBenchmark.assemble:·gc.alloc.rate.norm               avgt   10  362019.469 ±     23.645    B/op
[info] PhaseAssemblyBenchmark.assemble:·gc.churn.PS_Eden_Space           avgt   10     231.316 ±    132.517  MB/sec
[info] PhaseAssemblyBenchmark.assemble:·gc.churn.PS_Eden_Space.norm      avgt   10  330798.077 ± 188745.508    B/op
[info] PhaseAssemblyBenchmark.assemble:·gc.churn.PS_Survivor_Space       avgt   10       0.062 ±      0.140  MB/sec
[info] PhaseAssemblyBenchmark.assemble:·gc.churn.PS_Survivor_Space.norm  avgt   10      88.923 ±    201.403    B/op
[info] PhaseAssemblyBenchmark.assemble:·gc.count                         avgt   10      14.000               counts
[info] PhaseAssemblyBenchmark.assemble:·gc.time                          avgt   10      15.000                   ms
```