@retronym 2019-08-07
Okay, pushed a benchmark that shows the algorithmic problem. In it, each of the `size` phases declares a dependency to run after the two prior phases.

Before:
```
[info] Benchmark                        (size)  Mode  Cnt        Score       Error  Units
[info] PhaseAssemblyBenchmark.assemble       1  avgt    5     1509.672 ±   450.020  ns/op
[info] PhaseAssemblyBenchmark.assemble       4  avgt    5     5380.669 ±   854.188  ns/op
[info] PhaseAssemblyBenchmark.assemble       8  avgt    5    40450.627 ±  1995.082  ns/op
[info] PhaseAssemblyBenchmark.assemble      16  avgt    5  1255818.095 ± 82024.968  ns/op
```

This PR:
```
[info] Benchmark                        (size)  Mode  Cnt       Score       Error  Units
[info] PhaseAssemblyBenchmark.assemble       1  avgt    5    1003.871 ±    65.471  ns/op
[info] PhaseAssemblyBenchmark.assemble       4  avgt    5    2762.134 ±   193.734  ns/op
[info] PhaseAssemblyBenchmark.assemble       8  avgt    5   13104.495 ±   899.441  ns/op
[info] PhaseAssemblyBenchmark.assemble      16  avgt    5  481946.303 ± 30771.517  ns/op
```

I'll implement a better algorithm which should be proportional to number of phases + number of edges.